move git.go into ./vcs/git/git.go because another vcs(s) will be added.
